requirements.txt
adjustText==1.3.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
certifi==2025.4.26
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
fonttools==4.58.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
jiter==0.9.1
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
networkx==3.4.2
numpy==2.2.6
ollama==0.4.8
openai==1.79.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
yml==0.0.1
